+++ docs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Fixed
 * Fix timezone handling. Now a string the timezone name is saved
+* Hotfix: Fix session spanning in more than one day (due to worktime-py error)
 * Broken activity indicators in dashboard and subject show
 * Minor UI: Add air in work subject picker